txn: rename tm_tx_hash_ranges to tx_hash_ranges
