+++ src/v/cluster/tests/CMakeLists.txt
@@ -23,7 +23,7 @@ set(srcs
     feature_barrier_test.cc
     tm_stm_tests.cc
     tm_coordinator_mapper_tests.cc
-    tm_tx_hash_ranges_tests.cc
+    tx_hash_ranges_tests.cc
     rm_stm_tests.cc
     rm_stm_compatibility_test.cc
     id_allocator_stm_test.cc
