+++ src/v/cluster/tests/tx_hash_ranges_tests.cc
@@ -7,7 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-#include "cluster/tm_tx_hash_ranges.h"
+#include "cluster/tm_stm.h"
+#include "cluster/tx_hash_ranges.h"
 
 #include <seastar/testing/thread_test_case.hh>
 
@@ -22,7 +23,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_union_test) {
       cluster::tx_tm_hash_max - 99, cluster::tx_tm_hash_max);
     cluster::tx_hash_ranges_errc add_res;
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_1{};
+    cluster::tm_stm::hosted_txs hosted_transactions_1{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_1, range_start);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -38,7 +39,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_union_test) {
       hosted_transactions_1.hash_ranges.ranges[0].last,
       cluster::tx_tm_hash_max);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_2{};
+    cluster::tm_stm::hosted_txs hosted_transactions_2{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_2, range_start);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -54,7 +55,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_union_test) {
       hosted_transactions_2.hash_ranges.ranges[0].last,
       cluster::tx_tm_hash_max);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_3{};
+    cluster::tm_stm::hosted_txs hosted_transactions_3{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_3, range_end);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -78,7 +79,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_intersects_test) {
     cluster::tx_hash_range range_150_170(150, 170);
     cluster::tx_hash_range range_170_200(170, 200);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_1{};
+    cluster::tm_stm::hosted_txs hosted_transactions_1{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_1, range_100_200);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -92,7 +93,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_intersects_test) {
       hosted_transactions_1, range_170_200);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::intersection);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_2{};
+    cluster::tm_stm::hosted_txs hosted_transactions_2{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_2, range_100_150);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -103,7 +104,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_addition_intersects_test) {
       hosted_transactions_2, range_170_200);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_3{};
+    cluster::tm_stm::hosted_txs hosted_transactions_3{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_3, range_150_170);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -119,7 +120,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_include_exclude_test) {
     cluster::tx_hash_ranges_errc add_res;
 
     kafka::transactional_id tx_id("tx_1");
-    cluster::tm_tx_hosted_transactions hosted_transactions_1{};
+    cluster::tm_stm::hosted_txs hosted_transactions_1{};
     add_res = cluster::hosted_transactions::add_range(
       hosted_transactions_1, {0, cluster::tx_tm_hash_max});
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::success);
@@ -138,7 +139,7 @@ SEASTAR_THREAD_TEST_CASE(hash_ranges_include_exclude_test) {
       cluster::hosted_transactions::contains(hosted_transactions_1, tx_id));
     BOOST_REQUIRE_EQUAL(hosted_transactions_1.included_transactions.size(), 0);
 
-    cluster::tm_tx_hosted_transactions hosted_transactions_2{};
+    cluster::tm_stm::hosted_txs hosted_transactions_2{};
     add_res = cluster::hosted_transactions::exclude_transaction(
       hosted_transactions_2, tx_id);
     BOOST_REQUIRE(add_res == cluster::tx_hash_ranges_errc::not_hosted);
