+++ src/v/cluster/tm_stm.cc
@@ -10,7 +10,6 @@
 #include "cluster/tm_stm.h"
 
 #include "cluster/logger.h"
-#include "cluster/tm_tx_hash_ranges.h"
 #include "cluster/types.h"
 #include "model/record.h"
 #include "raft/errc.h"
@@ -99,8 +98,7 @@ model::record_batch tm_stm::serialize_tx(tm_transaction tx) {
     return do_serialize_tx(old_tx);
 }
 
-model::record_batch
-tm_stm::serialize_hosted_transactions(tm_tx_hosted_transactions hr) {
+model::record_batch tm_stm::serialize_hosted_transactions(hosted_txs hr) {
     storage::record_batch_builder b(
       model::record_batch_type::tx_tm_hosted_trasactions, model::offset(0));
     b.add_raw_kv(
@@ -143,7 +141,7 @@ ss::future<tm_stm::op_status> tm_stm::try_init_hosted_transactions(
     model::partition_id partition = get_partition();
     auto initial_hash_range = default_hash_range(
       partition, tx_coordinator_partition_amount);
-    tm_tx_hosted_transactions initial_hosted_transactions{};
+    hosted_txs initial_hosted_transactions{};
     auto res = hosted_transactions::add_range(
       initial_hosted_transactions, initial_hash_range);
     if (res == tx_hash_ranges_errc::success) {
@@ -328,14 +326,14 @@ tm_stm::do_sync(model::timeout_clock::duration timeout) {
     co_return _insync_term;
 }
 
-ss::future<tm_stm::op_status> tm_stm::update_hosted_transactions(
-  model::term_id term, tm_tx_hosted_transactions hr) {
+ss::future<tm_stm::op_status>
+tm_stm::update_hosted_transactions(model::term_id term, hosted_txs hr) {
     auto gh = _gate.hold();
     co_return co_await do_update_hosted_transactions(term, std::move(hr));
 }
 
-ss::future<tm_stm::op_status> tm_stm::do_update_hosted_transactions(
-  model::term_id term, tm_tx_hosted_transactions hr) {
+ss::future<tm_stm::op_status>
+tm_stm::do_update_hosted_transactions(model::term_id term, hosted_txs hr) {
     auto batch = serialize_hosted_transactions(std::move(hr));
 
     auto r = co_await replicate_quorum_ack(term, std::move(batch));
@@ -1017,8 +1015,7 @@ ss::future<> tm_stm::apply_hosted_transactions(model::record_batch b) {
       "{}",
       model::record_batch_type::tx_tm_hosted_trasactions,
       key);
-    auto hash_ranges = serde::from_iobuf<tm_tx_hosted_transactions>(
-      rec.release_value());
+    auto hash_ranges = serde::from_iobuf<hosted_txs>(rec.release_value());
     _hosted_txes = hash_ranges;
     return ss::now();
 }
