+++ src/v/cluster/tx_hash_ranges.h
@@ -153,25 +153,6 @@ struct tx_hash_ranges_set
 
 using repartitioning_id = named_type<int64_t, struct repartitioning_id_type>;
 
-struct draining_txs
-  : serde::envelope<draining_txs, serde::version<0>, serde::compat_version<0>> {
-    repartitioning_id id;
-    tx_hash_ranges_set ranges{};
-    absl::btree_set<kafka::transactional_id> transactions{};
-
-    draining_txs() = default;
-
-    draining_txs(
-      repartitioning_id id,
-      tx_hash_ranges_set ranges,
-      absl::btree_set<kafka::transactional_id> txs)
-      : id(id)
-      , ranges(std::move(ranges))
-      , transactions(std::move(txs)) {}
-
-    auto serde_fields() { return std::tie(id, ranges, transactions); }
-};
-
 namespace hosted_transactions {
 template<class T>
 tx_hash_ranges_errc
@@ -249,41 +230,6 @@ bool contains(T& self, const kafka::transactional_id& tx_id) {
 }
 } // namespace hosted_transactions
 
-struct tm_tx_hosted_transactions
-  : serde::envelope<
-      tm_tx_hosted_transactions,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    bool inited{false};
-    tx_hash_ranges_set hash_ranges{};
-    absl::btree_set<kafka::transactional_id> excluded_transactions{};
-    absl::btree_set<kafka::transactional_id> included_transactions{};
-    draining_txs draining{};
-
-    tm_tx_hosted_transactions() = default;
-
-    tm_tx_hosted_transactions(
-      bool inited,
-      tx_hash_ranges_set hr,
-      absl::btree_set<kafka::transactional_id> et,
-      absl::btree_set<kafka::transactional_id> it,
-      draining_txs dr)
-      : inited(inited)
-      , hash_ranges(std::move(hr))
-      , excluded_transactions(std::move(et))
-      , included_transactions(std::move(it))
-      , draining(std::move(dr)) {}
-
-    auto serde_fields() {
-        return std::tie(
-          inited,
-          hash_ranges,
-          excluded_transactions,
-          included_transactions,
-          draining);
-    }
-};
-
 } // namespace cluster
 
 namespace reflection {
@@ -310,53 +256,6 @@ struct adl<cluster::tx_hash_ranges_set> {
           = reflection::adl<std::vector<cluster::tx_hash_range>>{}.from(in);
         return {std::move(ranges)};
     }
-};
-
-template<>
-struct adl<cluster::draining_txs> {
-    void to(iobuf& out, cluster::draining_txs&& dr) {
-        reflection::serialize(out, dr.id, dr.ranges, dr.transactions);
-    }
-    cluster::draining_txs from(iobuf_parser& in) {
-        auto id = reflection::adl<cluster::repartitioning_id>{}.from(in);
-        auto ranges
-          = reflection::adl<std::vector<cluster::tx_hash_range>>{}.from(in);
-        auto txs = reflection::adl<absl::btree_set<kafka::transactional_id>>{}
-                     .from(in);
-        return {id, std::move(ranges), std::move(txs)};
-    }
-};
-
-template<>
-struct adl<cluster::tm_tx_hosted_transactions> {
-    void to(iobuf& out, cluster::tm_tx_hosted_transactions&& hr) {
-        reflection::serialize(
-          out,
-          hr.inited,
-          hr.hash_ranges,
-          hr.excluded_transactions,
-          hr.included_transactions,
-          hr.draining);
-    }
-    cluster::tm_tx_hosted_transactions from(iobuf_parser& in) {
-        auto inited = reflection::adl<bool>{}.from(in);
-
-        auto hash_ranges_set
-          = reflection::adl<cluster::tx_hash_ranges_set>{}.from(in);
-        auto included_transactions
-          = reflection::adl<absl::btree_set<kafka::transactional_id>>{}.from(
-            in);
-        auto excluded_transactions
-          = reflection::adl<absl::btree_set<kafka::transactional_id>>{}.from(
-            in);
-        auto draining = reflection::adl<cluster::draining_txs>{}.from(in);
-        return {
-          inited,
-          std::move(hash_ranges_set),
-          std::move(included_transactions),
-          std::move(excluded_transactions),
-          std::move(draining)};
-    };
 };
 
 } // namespace reflection
